requirements.txt
pytest
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
discord.py==1.3.4
Django==3.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
image==1.5.32
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
websockets==8.1
Werkzeug==1.0.1
yarl==1.5.1